requirements.txt
cramjam==2.8.3
fastparquet==2024.5.0
fsspec==2024.6.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
six==1.16.0
tzdata==2024.1
